EntryPoint
Since this code runs, this is the entry point to our entire project. Now its a matter of making sure we can run this and have it launch into the mGBA emulator, which will be rather important for basic windowing. Currently, that link isn't there, so we will need to figure out how to get things to be cohesive with regards to building and then having the mGBA emulator generate a window