EntryPoint
This is the bare minimum required to work with the hardware directly, and hence, is the minimum required for our application.

Successfully, we have a working build for my M2 Mac. This will allow me to write my first simplistic game, which has more to do with me wanting to work with Assembly code directly for graphics (since there aren't many convenient ways outside of stuff like this)